Remove old sidebar component
